+++ CHANGELOG.md
@@ -14,9 +14,13 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Added graduation rate and loan default rate external links
 - Changed the scale of the loan default rate graph to 40%
 - General code housekeeping (GitHub templates, linting improvements)
+- Added load programs script
 - loanDefaultRate values (three of them) removed from national-stats API calls
 - metric-view refactored
 - 10/25 loan term toggles added to metrics section
+- Include half year options for program length
+- Shift line item dollar signs over to accommodate six figure values
+- add `operating` field to admin
 
 ## 2.1.1
 - Parent PLUS loans separated from other family contributions